requirements.txt
discord.py
flask
uvicorn
fastapi